+++ mood-meter/src/App.js
@@ -3,7 +3,7 @@ import React from 'react';
 import './App.css';
 import Header from './components/Header.js'
 import Login from './components/Login.js'
-import Graph from './components/Graph'
+import Graph from './components/mood-graph/Graph'
 
 function App() {
   return (
