+++ mood-meter/src/components/mood-graph/moodBubble.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function MoodBubble() {
+    return (
+        <div className='bubble-button_container'>
+            <button className='bubble-button'>
+
+            </button>
+        </div>
+    )
+}
+export default MoodBubble
